# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Teacher {
  id: number;
  name: string;
  subject: string;
  maxWeeklyHours: number;
  weeklySubstitutionCount: number;
}

export interface CreateTeacher {
  name: string;
  subject: string;
  maxWeeklyHours: number;
}

export interface SchoolClass {
  id: number;
  grade: string;
  section: string;
}

export interface CreateClass {
  grade: string;
  section: string;
}

export interface Subject {
  id: number;
  name: string;
}

export interface CreateSubject {
  name: string;
}

export interface TimetableSlot {
  id: number;
  /** 0=Sunday, 1=Monday, ..., 4=Thursday */
  dayOfWeek: number;
  /** 1-8 */
  period: number;
  teacherId: number;
  classId: number;
  subjectId: number;
  teacherName: string;
  className: string;
  subjectName: string;
}

export interface CreateTimetableSlot {
  dayOfWeek: number;
  period: number;
  teacherId: number;
  classId: number;
  subjectId: number;
}

export interface SubstitutionAssignment {
  id: number;
  date: string;
  absentTeacherId: number;
  substituteTeacherId: number;
  timetableSlotId: number;
  absentTeacherName: string;
  substituteTeacherName: string;
  className: string;
  subjectName: string;
  period: number;
  dayOfWeek: number;
}

export interface DailySubstitutionPlan {
  date: string;
  absentTeacherIds: number[];
  assignments: SubstitutionAssignment[];
}

export interface AssignmentInput {
  timetableSlotId: number;
  substituteTeacherId: number;
}

export interface SaveSubstitutions {
  date: string;
  absentTeacherIds: number[];
  assignments: AssignmentInput[];
}

export interface RecommendSubstitutionsRequest {
  date: string;
  absentTeacherIds: number[];
}

export interface RecommendedTeacher {
  teacherId: number;
  teacherName: string;
  weeklySubstitutionCount: number;
  isFree: boolean;
}

export interface SubstitutionRecommendation {
  timetableSlotId: number;
  /** Period number */
  period: number;
  className: string;
  subjectName: string;
  absentTeacherName: string;
  recommendedTeachers: RecommendedTeacher[];
}

export interface WeeklySubstitutionCount {
  teacherId: number;
  teacherName: string;
  count: number;
}

export interface GenerateExamCommitteeRequest {
  examName: string;
  numberOfRooms: number;
  teacherIds: number[];
  proctorsPerRoom: number;
}

export interface RoomAssignment {
  roomNumber: number;
  teachers: Teacher[];
}

export interface SavedExamCommittee {
  id: number;
  examName: string;
  createdAt: string;
  assignments: RoomAssignment[];
}

export interface SaveExamCommitteeRequest {
  examName: string;
  assignments: RoomAssignment[];
}

export type GetSubstitutionsParams = {
  date: string;
};